server types migration
