+++ packages/server/admin/package.json
@@ -9,6 +9,8 @@
   "devDependencies": {
     "@uwdsc/eslint-config": "workspace:*",
     "@uwdsc/typescript-config": "workspace:*",
+    "@uwdsc/db": "workspace:*",
+    "@uwdsc/types": "workspace:*",
     "@types/node": "^20.19.9",
     "eslint": "^9.32.0",
     "typescript": "^5.9.2"
