--- packages/server/core/src/types/auth.ts
@@ -1,15 +1,0 @@
-export interface RegisterData {
-  email: string;
-  password: string;
-}
-
-export interface LoginData {
-  email: string;
-  password: string;
-}
-
-export interface RegisterCredentials {
-  email: string;
-  password: string;
-  emailRedirectTo?: string;
-}
