--- packages/server/core/src/types/errors.ts
@@ -1,10 +1,0 @@
-export class ApiError extends Error {
-  constructor(
-    message: string,
-    public statusCode: number = 500,
-    public code?: string,
-  ) {
-    super(message);
-    this.name = "ApiError";
-  }
-}
