+++ packages/types/src/entities/index.ts
@@ -1,1 +1,3 @@
 export * from "./file";
+export * from "./profile";
+export * from "./membership";
