+++ packages/types/src/entities/membership.ts
@@ -1,0 +1,5 @@
+export interface MarkAsPaidData {
+  payment_method: "cash" | "online" | "math_soc";
+  payment_location: string;
+  verifier: string;
+}
